requirements.txt
flask 
flask-pymongo
bcrypt
PyJWT
python-dotenv
